Bug 16444: Update audit process for Firefox ESR
The GENERAL document contains notes about a new procedure for reviewing
new features in Firefox releases for the next ESR Tor Browser will be
based on. The document is named GENERAL as we might want to describe
other parts of the audit process at some point, too, making this a proper
documentation of the audit process itself.